+++ web/src/components/core/empty/MqEmpty.tsx
