+++ web/src/routes/datasets/Datasets.tsx
@@ -10,7 +10,7 @@ import { encodeNode } from '../../helpers/nodes'
 import { fetchDatasets, resetDatasets } from '../../store/actionCreators'
 import { formatUpdatedAt } from '../../helpers'
 import Box from '@material-ui/core/Box'
-import MqEmpty from '../../components/core/empty/Empty'
+import MqEmpty from '../../components/core/empty/MqEmpty'
 import MqText from '../../components/core/text/MqText'
 import React from 'react'
 import createStyles from '@material-ui/core/styles/createStyles'
