+++ web/src/routes/jobs/Jobs.tsx
@@ -11,7 +11,7 @@ import { fetchJobs, resetJobs } from '../../store/actionCreators'
 import { formatUpdatedAt } from '../../helpers'
 import { stopWatchDuration } from '../../helpers/time'
 import Box from '@material-ui/core/Box'
-import MqEmpty from '../../components/core/empty/Empty'
+import MqEmpty from '../../components/core/empty/MqEmpty'
 import MqText from '../../components/core/text/MqText'
 import React from 'react'
 import createStyles from '@material-ui/core/styles/createStyles'
